requirements.txt
aiofiles==25.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
ddddocr==1.6.1
et_xmlfile==2.0.0
fastapi==0.129.0
flake8==7.3.0
flatbuffers==25.12.19
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
magic-filter==1.0.12
mccabe==0.7.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
onnxruntime==1.24.4
opencv-python==4.13.0.92
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pefile==2024.8.26
pillow==12.2.0
playwright==1.58.0
playwright-stealth==2.0.3
propcache==0.4.1
protobuf==7.34.1
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pyflakes==3.4.0
PyGetWindow==0.0.9
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyMsgBox==2.0.1
pyperclip==1.11.0
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6_sip==13.11.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.32.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.5
selenium==4.40.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.52.1
sympy==1.14.0
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
